# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DOSTAC website + admin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ok {
  ok: boolean;
}

export interface ApiError {
  error: string;
}

export type Lang = (typeof Lang)[keyof typeof Lang];

export const Lang = {
  ko: "ko",
  en: "en",
  ja: "ja",
  zh: "zh",
  vi: "vi",
} as const;

export interface Translation {
  lang: Lang;
  name: string;
  headline: string;
  body: string;
}

export interface NoticeTranslation {
  lang: Lang;
  title: string;
  excerpt: string;
  body: string;
}

export interface PublicProduct {
  id: number;
  slug: string;
  category: string;
  sortOrder: number;
  /** @nullable */
  imageUrl?: string | null;
  name: string;
  headline: string;
  body: string;
}

export interface PublicNotice {
  id: number;
  slug: string;
  category: string;
  region: string;
  publishedAt: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  title: string;
  excerpt: string;
  body: string;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
  role: string;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminProduct {
  id: number;
  slug: string;
  category: string;
  sortOrder: number;
  /** @nullable */
  imageUrl?: string | null;
  published: boolean;
  translations: Translation[];
}

export interface AdminProductInput {
  slug: string;
  category: string;
  sortOrder: number;
  /** @nullable */
  imageUrl?: string | null;
  published: boolean;
  translations: Translation[];
}

export interface AdminNotice {
  id: number;
  slug: string;
  category: string;
  region: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  published: boolean;
  publishedAt: string;
  translations: NoticeTranslation[];
}

export interface AdminNoticeInput {
  slug: string;
  category: string;
  region: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  published: boolean;
  publishedAt: string;
  translations: NoticeTranslation[];
}

export interface Inquiry {
  id: number;
  company: string;
  name: string;
  email: string;
  phone: string;
  country: string;
  projectType: string;
  productInterest: string[];
  monthlyVolume: string;
  message: string;
  status: string;
  adminNote: string;
  source: string;
  createdAt: string;
}

export type InquirySummaryByStatus = {
  new: number;
  in_progress: number;
  done: number;
};

export interface InquirySummary {
  total: number;
  byStatus: InquirySummaryByStatus;
  recent: Inquiry[];
}

export interface InquiryUpdate {
  status?: string;
  adminNote?: string;
}

export interface ContactInquiryInput {
  company: string;
  name: string;
  email: string;
  phone?: string;
  country?: string;
  projectType?: string;
  productInterest?: string[];
  monthlyVolume?: string;
  message: string;
}

export type TranslateInputFormat =
  (typeof TranslateInputFormat)[keyof typeof TranslateInputFormat];

export const TranslateInputFormat = {
  text: "text",
  html: "html",
} as const;

export interface TranslateInput {
  sourceText: string;
  sourceLang?: Lang;
  targetLangs: Lang[];
  /** Hint about the field — e.g. "product name", "marketing headline", "rich html body" */
  context?: string;
  format?: TranslateInputFormat;
}

export type TranslateResultTranslationsItem = {
  lang: Lang;
  text: string;
};

export interface TranslateResult {
  translations: TranslateResultTranslationsItem[];
}

export interface UploadSignInput {
  filename: string;
  contentType: string;
}

export interface UploadSignResult {
  uploadUrl: string;
  publicUrl: string;
  key: string;
}

/**
 * Not found
 */
export type NotFoundResponse = ApiError;

/**
 * Unauthorized
 */
export type UnauthorizedResponse = ApiError;

/**
 * Bad request
 */
export type BadRequestResponse = ApiError;

export type ListPublicProductsParams = {
  lang: Lang;
  category?: string;
};

export type GetPublicProductParams = {
  lang: Lang;
};

export type ListPublicNoticesParams = {
  lang: Lang;
  category?: string;
  search?: string;
};

export type GetPublicNoticeParams = {
  lang: Lang;
};

export type AdminListInquiriesParams = {
  status?: string;
};